adding new user to the room

+++ API_src/src/main/java/pl/azurix/user/UserRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
     @Query("SELECT t.id FROM User t WHERE t.login=:login AND t.password=:password")
-    Integer login(@Param("login") String login, @Param("password") String password);
+    Long login(@Param("login") String login, @Param("password") String password);
 }
